requirements.txt
pygame==1.9.3
PyOpenGL==3.1.0
Pillow==4.2.1
bitstring==3.1.5
